Search all headers for WCS data, not just the first one.
Previously only the first header would be loaded and checked for WCS data, but many images put this information in an extension. The upload functionality now iterates over all headers in the file and uses the data from the first one with valid WCS.

The new functionality relies on the HTTP server supporting requests for particular ranges within the files. Not all servers allow this; if it's not supported, we fall back to using the primary header only. The only alternative I can think of would be to download the entire file.